# 01-coupling/03-loosely-coupled-generated/internal/http_types.go
// Package internal provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.2 DO NOT EDIT.
package internal

// EmailResponse defines model for EmailResponse.
type EmailResponse struct {
	Address string `json:"address"`
	Primary bool   `json:"primary"`
}

// PatchUserRequest defines model for PatchUserRequest.
type PatchUserRequest struct {
	// First name
	FirstName *string `json:"first_name,omitempty"`

	// Last name
	LastName *string `json:"last_name,omitempty"`
}

// PostUserRequest defines model for PostUserRequest.
type PostUserRequest struct {
	// E-mail
	Email string `json:"email"`

	// First name
	FirstName string `json:"first_name"`

	// Last name
	LastName string `json:"last_name"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	DisplayName string          `json:"display_name"`
	Emails      []EmailResponse `json:"emails"`
	FirstName   string          `json:"first_name"`
	Id          int             `json:"id"`
	LastName    string          `json:"last_name"`
}

// UsersResponse defines model for UsersResponse.
type UsersResponse []UserResponse

// UserID defines model for userID.
type UserID string

// PostUserJSONBody defines parameters for PostUser.
type PostUserJSONBody PostUserRequest

// PatchUserJSONBody defines parameters for PatchUser.
type PatchUserJSONBody PatchUserRequest

// PostUserJSONRequestBody defines body for PostUser for application/json ContentType.
type PostUserJSONRequestBody PostUserJSONBody

// PatchUserJSONRequestBody defines body for PatchUser for application/json ContentType.
type PatchUserJSONRequestBody PatchUserJSONBody